Refactor config gui extensions
